pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "wqe"
version = "2024.05.30"
requires-python = ">=3.10"
authors = [
    {name = "Artur Kulmizev", email = "artur.kulmizev@kuleuven.be"},
    {name = "Kushal Tatariya", email = "kushal.tatariya@kuleuven.be"}
]
maintainers = [
    {name = "Artur Kulmizev", email = "artur.kulmizev@kuleuven.be"}
]
description = "A package for loading and filtering Wikipedia data."
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent"]
dependencies = [
    "hydra-core",
    "datasets",
    "tokenizers",
    "transformers",
    "accelerate",
    "huggingface_hub",
    "fasttext",
    "nltk",
    "scikit-learn",
    "wandb",
    "tqdm"
]

[project.urls]
"Homepage" = "https://github.com/kushaltatariya/Wikipedia_Quality_Estimation"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"wqe.data" = ["resources/wiki_mappings.json"]